requirements.txt
Pillow
imagehash
imutils
opencv-python
Numpy
tqdm
scipy
numpy
piexif
ocrd-pyexiftool
ZODB